Create wallets/lib/crypto.js
